Refactor exercise form components
- Moved ExerciseForm and related components to a new `exercises` module for better organization.
- Extracted basic fields, form actions, type change confirmation, and type-specific fields into separate components.
- Updated references in ExerciseManager and ExerciseDetail to use the new module structure.
- Improved code readability and maintainability by modularizing the exercise form logic.
