+++ src/components/exercise_manager.rs
@@ -1,4 +1,4 @@
-use crate::components::exercise_form::{ExerciseForm, FormMode};
+use crate::components::exercises::{ExerciseForm, FormMode};
 use crate::models::{exercise::Exercise, repository::{get_exercise_repository, ExerciseRepository}};
 use leptos::prelude::*;
