+++ src/components/exercises/exercise_type_specific_fields.rs
@@ -1,0 +1,137 @@
+use leptos::prelude::*;
+use crate::components::{
+    svg_fretboard_scale_display::SvgFretboardScaleDisplay,
+};
+use super::PositionPresetButtons;
+use crate::music::{heptatonic_scales::HeptaScaleType, notes::Note, scales::ScaleType};
+
+#[component]
+pub fn ExerciseTypeSpecificFields(
+    exercise_type: ReadSignal<String>,
+    root_note: ReadSignal<Note>,
+    on_root_note_change: Callback<Note>,
+    scale_type: ReadSignal<ScaleType>, 
+    on_scale_type_change: Callback<ScaleType>,
+    min_fret: ReadSignal<u8>,
+    on_min_fret_change: Callback<u8>,
+    max_fret: ReadSignal<u8>,
+    on_max_fret_change: Callback<u8>,
+) -> impl IntoView {
+    // Handle preset selection
+    let on_preset_select = Callback::new(move |(min, max): (u8, u8)| {
+        on_min_fret_change.run(min);
+        on_max_fret_change.run(max);
+    });
+
+    view! {
+        // Conditional fields for Scale and Triad types
+        {move || {
+            let ex_type = exercise_type.get();
+            if ex_type == "Scale" || ex_type == "Triad" {
+                view! {
+                    <div class="space-y-4 p-4 bg-gray-50 rounded-md">
+                        <div class="grid grid-cols-1 md:grid-cols-2 gap-4">
+                            // Root note selector
+                            <div>
+                                <label class="block text-sm font-medium text-gray-700 mb-1">Root Note</label>
+                                <select
+                                    class="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
+                                    prop:value={move || format!("{:?}", root_note.get())}
+                                    on:change=move |e| {
+                                        let value = event_target_value(&e);
+                                        if let Ok(note) = value.parse::<Note>() {
+                                            on_root_note_change.run(note);
+                                        }
+                                    }
+                                >
+                                    <option value="C">C</option>
+                                    <option value="Cs">"C#"</option>
+                                    <option value="D">D</option>
+                                    <option value="Ds">"D#"</option>
+                                    <option value="E">E</option>
+                                    <option value="F">F</option>
+                                    <option value="Fs">"F#"</option>
+                                    <option value="G">G</option>
+                                    <option value="Gs">"G#"</option>
+                                    <option value="A">A</option>
+                                    <option value="As">"A#"</option>
+                                    <option value="B">B</option>
+                                </select>
+                            </div>
+
+                            // Scale type selector (simplified for now)
+                            <div>
+                                <label class="block text-sm font-medium text-gray-700 mb-1">Scale Type</label>
+                                <select
+                                    class="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
+                                    on:change=move |e| {
+                                        let value = event_target_value(&e);
+                                        match value.as_str() {
+                                            "Major" => on_scale_type_change.run(ScaleType::Hepatonic(HeptaScaleType::Major)),
+                                            "Minor" => on_scale_type_change.run(ScaleType::Hepatonic(HeptaScaleType::Minor)),
+                                            _ => on_scale_type_change.run(ScaleType::Hepatonic(HeptaScaleType::Major)),
+                                        }
+                                    }
+                                >
+                                    <option value="Major">Major</option>
+                                    <option value="Minor">Natural Minor</option>
+                                </select>
+                            </div>
+                        </div>
+
+                        // Fret range
+                        <div class="grid grid-cols-1 md:grid-cols-2 gap-4">
+                            <div>
+                                <label class="block text-sm font-medium text-gray-700 mb-1">Min Fret</label>
+                                <input
+                                    type="number"
+                                    min="0"
+                                    max="24"
+                                    class="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
+                                    prop:value={move || min_fret.get().to_string()}
+                                    on:input=move |e| {
+                                        if let Ok(val) = event_target_value(&e).parse::<u8>() {
+                                            on_min_fret_change.run(val.min(24));
+                                        }
+                                    }
+                                />
+                            </div>
+                            <div>
+                                <label class="block text-sm font-medium text-gray-700 mb-1">Max Fret</label>
+                                <input
+                                    type="number"
+                                    min="0"
+                                    max="24"
+                                    class="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
+                                    prop:value={move || max_fret.get().to_string()}
+                                    on:input=move |e| {
+                                        if let Ok(val) = event_target_value(&e).parse::<u8>() {
+                                            on_max_fret_change.run(val.min(24));
+                                        }
+                                    }
+                                />
+                            </div>
+                        </div>
+
+                        // Position presets
+                        <PositionPresetButtons on_preset_select />
+
+                        // Fretboard preview for scales
+                        <div>
+                            <label class="block text-sm font-medium text-gray-700 mb-2">Preview</label>
+                            <div class="bg-gray-50 rounded-lg p-4 max-w-2xl mx-auto">
+                                <SvgFretboardScaleDisplay
+                                    fret_range={Signal::derive(move || min_fret.get() as usize..=max_fret.get() as usize)}
+                                    root_note={Signal::derive(move || root_note.get())}
+                                    scale_type={Signal::derive(move || scale_type.get())}
+                                />
+                            </div>
+                        </div>
+                    </div>
+                }.into_any()
+            } else {
+                view! { <div></div> }.into_any()
+            }
+        }}
+    }
+}
