+++ src/components/mod.rs
@@ -1,4 +1,4 @@
-pub mod exercise_form;
+pub mod exercises;
 pub mod exercise_manager;
 pub mod fret_range_selector;
 pub mod fretboard;
